Refactor metric submissions back into check
