+++ kafka_consumer/datadog_checks/kafka_consumer/client/confluent_kafka_client.py
@@ -11,11 +11,5 @@ class ConfluentKafkaClient:
     def get_broker_offset(self):
         pass
 
-    def report_consumer_offset_and_lag(self):
-        pass
-
-    def report_broker_offset(self):
-        pass
-
     def collect_broker_metadata(self):
         pass
