+++ kafka_consumer/datadog_checks/kafka_consumer/client/kafka_client.py
@@ -25,13 +25,5 @@ class KafkaClient(ABC):
         pass
 
     @abstractmethod
-    def report_consumer_offsets_and_lag(self):
-        pass
-
-    @abstractmethod
-    def report_highwater_offsets(self):
-        pass
-
-    @abstractmethod
     def collect_broker_metadata(self):
         pass
